requirements.txt
asgiref==3.8.1
Django==5.0.7
djangorestframework==3.15.2
gunicorn==22.0.0
packaging==24.1
sqlparse==0.5.0
tzdata==2024.1
